VFB-42 First commit. Make new common folder
