+++ src/app/clients/edit/[id]/page.tsx
@@ -1,10 +1,11 @@
 import { Metadata } from "next";
 
 import React from "react";
+import supabase from "@/supabaseServer";
 import ClientForm from "@/app/clients/add/ClientForm";
 import { Errors, FormErrors } from "@/components/Form/formFunctions";
 import autofill from "@/app/clients/edit/[id]/autofill";
-import { fetchClients, fetchFamilies } from "@/pdf/ShoppingList/databaseFetch";
+import { fetchClients, fetchFamilies } from "@/common/fetch";
 
 interface EditClientsParameters {
     params: { id: string };
@@ -13,8 +14,8 @@ interface EditClientsParameters {
 const EditClients: ({ params }: EditClientsParameters) => Promise<React.ReactElement> = async ({
     params,
 }: EditClientsParameters) => {
-    const clientData = await fetchClients(params.id);
-    const familyData = await fetchFamilies(clientData.family_id);
+    const clientData = await fetchClients(params.id, supabase);
+    const familyData = await fetchFamilies(clientData.family_id, supabase);
     const initialFields = autofill(clientData, familyData);
     const initialFormErrors: FormErrors = {
         fullName: Errors.none,
