+++ src/common/format.ts
@@ -1,47 +1,9 @@
-import {
-    fetchClients,
-    fetchComment,
-    fetchFamilies,
-    fetchLists,
-    fetchParcels,
-} from "@/pdf/ShoppingList/databaseFetch";
 import { Schema } from "@/database_utils";
 import { Person } from "@/components/Form/formFunctions";
-
-export interface BlockProps {
-    [key: string]: string;
-}
-
-interface ParcelInfoAndClientID {
-    parcelInfo: ParcelInfo;
-    clientID: string;
-}
-
-interface ClientDataAndFamilyData {
-    clientData: Schema["clients"];
-    familyData: Schema["families"][];
-}
 
 interface NappySizeAndExtraInformation {
     nappySize: string;
     extraInformation: string;
-}
-
-export interface ShoppingListPDFDataProps {
-    postcode: string;
-    parcelInfo: ParcelInfo;
-    clientSummary: ClientSummary;
-    householdSummary: HouseholdSummary;
-    requirementSummary: RequirementSummary;
-    itemsList: Item[];
-    endNotes: string;
-}
-
-export interface ParcelInfo extends BlockProps {
-    voucherNumber: string;
-    packingDate: string;
-    collectionDate: string;
-    collectionSite: string;
 }
 
 export interface ClientSummary {
@@ -51,14 +13,14 @@ export interface ClientSummary {
     extraInformation: string;
 }
 
-export interface HouseholdSummary extends BlockProps {
+export interface HouseholdSummary {
     householdSize: string;
     genderBreakdown: string;
     numberOfBabies: string;
     ageAndGenderOfChildren: string;
 }
 
-export interface RequirementSummary extends BlockProps {
+export interface RequirementSummary {
     feminineProductsRequired: string;
     babyProductsRequired: string;
     petFoodRequired: string;
@@ -66,51 +28,9 @@ export interface RequirementSummary extends BlockProps {
     otherItems: string;
 }
 
-export interface Item {
-    description: string;
-    quantity: string;
-    notes: string;
-}
-
 export const formatCamelCaseKey = (objectKey: string): string => {
     const withSpace = objectKey.replaceAll(/([a-z])([A-Z])/g, "$1 $2");
     return withSpace.charAt(0).toUpperCase() + withSpace.slice(1);
-};
-
-const formatDate = (dateString: string | null): string => {
-    if (!dateString) {
-        return "";
-    }
-    return new Date(dateString).toLocaleString("en-GB", {
-        year: "numeric",
-        month: "numeric",
-        day: "numeric",
-        hour: "numeric",
-        minute: "numeric",
-    });
-};
-
-const prepareParcelInfo = async (parcelID: string): Promise<ParcelInfoAndClientID> => {
-    const fetchedData = await fetchParcels(parcelID);
-    const parcelInfo: ParcelInfo = {
-        voucherNumber: fetchedData.voucher_number ?? "",
-        packingDate: formatDate(fetchedData.packing_datetime),
-        collectionDate: formatDate(fetchedData.collection_datetime),
-        collectionSite: fetchedData.collection_centre ?? "",
-    };
-
-    if (parcelInfo.collectionSite === "Delivery") {
-        parcelInfo.collectionSite = "N/A - Delivery";
-    }
-
-    return { parcelInfo: parcelInfo, clientID: fetchedData.client_id };
-};
-
-const getClientAndFamilyData = async (clientID: string): Promise<ClientDataAndFamilyData> => {
-    const clientData = await fetchClients(clientID);
-    const familyData = await fetchFamilies(clientData.family_id);
-
-    return { clientData: clientData, familyData: familyData };
 };
 
 export const processExtraInformation = (original: string): NappySizeAndExtraInformation => {
@@ -197,55 +117,3 @@ export const prepareRequirementSummary = (clientData: Schema["clients"]): Requir
         otherItems: displayList(clientData.other_items),
     };
 };
-
-const getQuantityAndNotes = (
-    row: Schema["lists"],
-    size: number
-): Pick<Item, "quantity" | "notes"> => {
-    if (size >= 10) {
-        size = 10;
-    }
-    const size_quantity = `${size}_quantity` as keyof Schema["lists"];
-    const size_notes = `${size}_notes` as keyof Schema["lists"];
-    return {
-        quantity: row[size_quantity]?.toString() ?? "",
-        notes: row[size_notes]?.toString() ?? "",
-    };
-};
-
-const prepareItemsList = async (householdSize: number): Promise<Item[]> => {
-    const listData = await fetchLists();
-    return listData.map((row): Item => {
-        return {
-            description: row.item_name,
-            ...getQuantityAndNotes(row, householdSize),
-        };
-    });
-};
-
-const prepareData = async (parcelID: string): Promise<ShoppingListPDFDataProps> => {
-    const { parcelInfo, clientID } = await prepareParcelInfo(parcelID);
-    const { clientData, familyData } = await getClientAndFamilyData(clientID);
-    const itemsList = await prepareItemsList(familyData.length);
-
-    const clientSummary = prepareClientSummary(clientData);
-    const householdSummary = prepareHouseholdSummary(familyData);
-    const requirementSummary = prepareRequirementSummary(clientData);
-
-    const { nappySize } = processExtraInformation(clientData.extra_information);
-    requirementSummary.babyProductsRequired += ` (${nappySize})`;
-
-    const endNotes = await fetchComment();
-
-    return {
-        postcode: clientData.address_postcode,
-        parcelInfo: parcelInfo,
-        clientSummary: clientSummary,
-        householdSummary: householdSummary,
-        requirementSummary: requirementSummary,
-        itemsList: itemsList,
-        endNotes: endNotes,
-    };
-};
-
-export default prepareData;
